feat: RoomRes 객체 생성 \n 채팅방 조회시 RoomRes객체 반환하도록 수정

+++ backend/src/main/java/com/ssafy/api/request/RoomCreatePostRequest.java
@@ -27,10 +27,10 @@ public class RoomCreatePostRequest {
     int limit;
 
     @ApiModelProperty(name="책 카테고리", example="0" )
-    int bookCategory;
+    Long bookCategory;
 
     @ApiModelProperty(name="영화 카테고리", example="1" )
-    int movieCategory;
+    Long movieCategory;
 
     @ApiModelProperty(name="비밀번호", example="1234dfsf" )
     String password;
